+++ ui/root/build.gradle.kts
@@ -14,6 +14,11 @@ kotlin {
                 implementation(projects.ui.common)
                 implementation(projects.ui.signedIn)
                 implementation(projects.ui.signedOut)
+                implementation(projects.domain.authentication)
+                implementation(compose.runtime)
+                implementation(compose.foundation)
+                implementation(compose.material)
+                implementation(libs.io.insert.koin.core)
             }
         }
