+++ ui/root/src/commonMain/kotlin/social/androiddev/root/navigation/DefaultRootComponent.kt
@@ -22,6 +22,7 @@ import com.arkivanov.decompose.value.Value
 import com.arkivanov.essenty.parcelable.Parcelable
 import com.arkivanov.essenty.parcelable.Parcelize
 import social.androiddev.root.navigation.DefaultRootComponent.Config
+import social.androiddev.root.splash.DefaultSplashComponent
 import social.androiddev.signedin.navigation.DefaultSignedInRootComponent
 import social.androiddev.signedout.root.DefaultSignedOutRootComponent
 import kotlin.coroutines.CoroutineContext
@@ -78,6 +79,7 @@ class DefaultRootComponent(
         componentContext: ComponentContext,
     ) = DefaultSplashComponent(
         componentContext = componentContext,
+        mainContext = mainContext,
         navigateToTimelineInternal = {
             navigation.replaceCurrent(Config.SignedIn)
         },
